refactor(query): split select_executor.rs into concern submodules (#1121)
Behavior-preserving module split (epic #1116). The 4060-line
select_executor.rs becomes select_executor/mod.rs (SelectExecutor
orchestration + entry points) plus per-concern submodules:

- value_ops.rs    value comparison + arithmetic primitives
- predicate.rs    SSTable leaf-predicate eval (pub evaluate_*/LeafOutcome)
- lookup.rs       partition/clustering lookup classification
- aggregation.rs  GROUP BY accumulation
- row_build.rs    scan-row assembly (pub build_row_from_scan)
- writetime_ttl.rs WRITETIME/TTL projection + injectable clock
- test_support.rs shared #[cfg(test)] fixtures (pub(crate))

mod.rs keeps the public surface via pub use re-exports, so
query::mod's re-export list and all external callers (cqlite-cli,
cqlite-flight, prepared.rs, engine.rs) resolve unchanged. Tests follow
their code into per-concern test modules; all 40 test fns preserved
(no count change, no rewrite). Cross-submodule helpers use pub(super);
public items keep pub. No semantic changes.

Low-risk decomposition only: introduced column_info_from_type_str to
dedupe the three identical ColumnInfo-from-schema build sites. The
mega async methods (execute, execute_streaming_background,
execute_sstable_scan) were relocated intact (decomposing them is
high-risk for a behavior-preserving change).
